Install HtmlSanitizer, edit RiotApi DTOs, finish ReturnChampionsData method in RiotSharpService

+++ Services/LoLStats.Services/Models/RiotApiDtos/RiotApiAbilityPerLevelDto.cs
@@ -2,14 +2,14 @@
 {
     public class RiotApiAbilityPerLevelDto
     {
-        public int Level { get; set; }
+        public byte Level { get; set; }
 
-        public int Cooldown { get; set; }
+        public double Cooldown { get; set; }
 
-        public int Cost { get; set; }
+        public double Cost { get; set; }
 
         public bool CostsPerSecond { get; set; }
 
-        public int Range { get; set; }
+        public double Range { get; set; }
     }
 }
